+++ src/utils.rs
@@ -16,10 +16,10 @@ pub fn write_ascii<W: Write>(dst: &mut W, byte: u8) -> std::io::Result<usize> {
 }
 
 #[inline]
-pub fn write_utf8<W: Write>(dst: &mut W, u: u16) -> std::io::Result<usize> {
+pub fn write_utf8<W: Write>(dst: &mut W, u: u32) -> std::io::Result<usize> {
     if u == 0 {
         write_decode_fallback(dst)
-    } else if let Some(ch) = char::from_u32(u as u32) {
+    } else if let Some(ch) = char::from_u32(u) {
         let mut buf = [0u8; 4];
         let s = ch.encode_utf8(&mut buf);
         dst.write_all(s.as_bytes())?;
